Code documentation
Summary:
The code prototype were considered as self-documented, using clear
variable and function names, dividing the problem to solve into
small functions.

We add documentation to add some notes and provide a foundation for
the future evolution of the code.

In addition, PEP 0257 convention is followed.

Test Plan: Proofread the code comments.

Reviewers: dereckson

Differential Revision: http://devcentral.nasqueron.org/D299