Declare notification exchange as durable
Summary:
When the notifications exchange is created on RabbitMQ by our container
setup script, it receives a Durable: true feature. We mirror this
behavior when we declare the exchange.

This fixes the following issue:

Channel error: (406, "PRECONDITION_FAILED - inequivalent arg 'durable'
for exchange 'notifications' in vhost 'dev': received 'false'
but current is 'true'")

Test Plan: Run the client, check we're well connected to the broker

Reviewers: dereckson

Differential Revision: http://devcentral.nasqueron.org/D335